dnssec-enabled code imported from poslib cvs (dnssec branch)

+++ poslib/exception.cpp
@@ -33,10 +33,10 @@ PException::PException(const char *_message) {
 }
 
 PException::PException(const char *_message, PException &p) {
-  int len = strlen(_message) + strlen(p.message) + 1;
-  message = (char *)malloc(len);
-  strncpy(message, _message, len-1);
-  strncat(message, p.message, len-1);
+  size_t len = strlen(_message) + strlen(p.message) + 1;
+  message = (char*) malloc(len);
+  strncpy(message, _message, len - 1);
+  strncat(message, p.message, len - 1);
 }
 
 PException::PException(const PException& p) {
